+++ src/CategoriesCollection.php
@@ -14,7 +14,7 @@ final class CategoriesCollection extends BaseCollection implements XmlConvertibl
 {
     protected string $class = Category::class;
 
-    public function appendToXml(\SimpleXMLElement &$parent): void
+    public function appendToXml(\SimpleXMLElement $parent): void
     {
         // @phpstan-ignore argument.type
         array_walk($this->items, static function (Category $value) use ($parent): void {