+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 0.14.0-dev
 - added RSS extension CreativeCommons
 - BC break: pubDate in items now has to be a DateTime
+- added support for isPermaLink for item's guid
 
 Version 0.13.0
 - raised minimal version of PHP to 8.3